add codec options for auto tag handling with (Un)Marshaler types
Per default, a type implementing the Marshaler or Unmarshaler interfaces
is expected to encode or decode CBOR tags itself, even if the type is
registered to a TagSet. Setting the new HandleTagForMarshaler encoder
option or HandleTagForUnmarshaler decoder option to true activates
automatic tag handling for Marshaler/Unmarshaler types in the same way
that this is implemented for all other types.